refactor model_helper.py, update fastapi with new sentiment score format

+++ model/api/main.py
@@ -25,7 +25,7 @@ class PredictionOutput(BaseModel):
     event_type: str 
 
 class SentimentOutput(BaseModel): 
-    sentiment_score: int
+    sentiment_score: float
 
 @app.get("/")
 def read_root():
